feat: add HamburgerMenu, HangmanDrawing, and HangmanGame components
- Implemented HamburgerMenu for language selection and dark mode toggle.
- Created HangmanDrawing to visually represent the hangman game state.
- Developed HangmanGame to manage game logic, including word selection and scoring.

feat: introduce PresetsModal and URLGenerator components

- Added PresetsModal for selecting word categories and managing word pairs.
- Implemented URLGenerator for generating game links with selected words and languages.

chore: add i18n support for localization

- Created i18n.js for managing translations in English and Finnish.
