+++ src/i18n.js
@@ -1,0 +1,90 @@
+const i18n = {
+  en: {
+    generateLink: "Generate Link",
+    addWord: "Add Word",
+    learningWordPlaceholder: "Learning language word",
+    nativeWordPlaceholder: "Native translation",
+    noWords: "No words configured!",
+    meaning: "Meaning",
+    hangmanDrawing: "Hangman Drawing",
+    wrongGuesses: "Wrong guesses",
+    guessed: "You guessed it!",
+    answerWas: "The word was",
+    nextWord: "Next Word",
+    score: "Score",
+    correct: "✅",
+    incorrect: "❌",
+    generateGameLinkTitle: "Generate Game Link",
+    selectLanguage: "Select Language",
+    congratsTitle: "Congratulations!",
+    congratsMsg: "You've guessed all words correctly at least once!",
+    closeStr: "Close",
+    restart: "Restart",
+    swapModeLabel: "Occasionally swap question/answer direction",
+    learningLangLabel: "Learning language",
+    nativeLangLabel: "Native language",
+    addCommonWords: "Add 5 common words",
+    newGame: "New Game",
+    presets: "Presets",
+    selectCategories: "Select categories",
+    howManyToAdd: "How many to add:",
+    addSelected: "Add selected",
+    unexclude: "Unexclude",
+    exclude: "Exclude",
+    word: "word",
+    s: "s",
+    match: "match",
+    the: "the",
+    selected: "selected",
+    categories: "categories",
+    removeWord: "Remove word",
+    removeIcon: "×",
+    presetsMatchCount: "{count} matching words in selected categories.",
+    removeAllWords: "Remove all words"
+  },
+  fi: {
+    generateLink: "Luo linkki",
+    addWord: "Lisää sana",
+    learningWordPlaceholder: "Opittava sana",
+    nativeWordPlaceholder: "Käännös omalla kielellä",
+    noWords: "Ei sanoja määritetty!",
+    meaning: "Merkitys",
+    hangmanDrawing: "Hirsipuupiirros",
+    wrongGuesses: "Väärät arvaukset",
+    guessed: "Arvasit oikein!",
+    answerWas: "Sana oli",
+    nextWord: "Seuraava sana",
+    score: "Pisteet",
+    correct: "✅",
+    incorrect: "❌",
+    generateGameLinkTitle: "Luo pelilinkki",
+    selectLanguage: "Valitse kieli",
+    congratsTitle: "Onnittelut!",
+    congratsMsg: "Arvasit kaikki sanat oikein!",
+    restart: "Aloita alusta",
+    swapModeLabel: "Vaihda kysymyksen ja vastauksen suuntaa satunnaisesti",
+    learningLangLabel: "Opittava kieli",
+    nativeLangLabel: "Oma kieli",
+    addCommonWords: "Lisää 5 yleistä sanaa",
+    newGame: "Uusi peli",
+    presets: "Oletukset",
+    selectCategories: "Valitse kategoriat",
+    howManyToAdd: "Kuinka monta lisätään:",
+    addSelected: "Lisää valitut",
+    closeStr: "Sulje",
+    unexclude: "Poista sulku",
+    exclude: "Sulje pois",
+    word: "sana",
+    s: "t",
+    match: "ottelu",
+    the: "se",
+    selected: "valitut",
+    categories: "kategoriat",
+    removeWord: "Poista sana",
+    removeIcon: "×",
+    presetsMatchCount: "{count} sanaa valituissa kategorioissa.",
+    removeAllWords: "Poista kaikki sanat"
+  }
+};
+
+export default i18n;
